Refactor to staticmethod classes holding each GLM family
